Crate saml_rs

A library for doing SAML things, terribly, in rust.

My main aim at the moment is to provide IdP capabilities for the Kanidm project.

If you would like to help - please log PRs/Issues against terminaloutcomes/saml-rs.

Modules

metadata
Handy for the XML metadata part of SAML

response
Want to build a SAML response? Here’s your module. 🥳

test_samples
Random samples of XML I’ve found around the place

xmlutils
Internal utilities for doing things with XML

Structs

AuthnDecodeError

SamlAuthnRequest
Stores the values one would expect in an AuthN Request

SamlAuthnRequestParser
Used to pull apart a SAML AuthN Request and build a SamlAuthnRequest

SamlQuery
Used in the SAML Redirect GET request to pull out the query values

Functions

_get_private_key

decode_authn_request_base64_encoded

parse_authn_request
Give it a string full of XML and it’ll give you back a SamlAuthnRequest object which has the details

parse_authn_tokenizer_attribute
Used inside SamlAuthnRequestParser to help parse the AuthN request

parse_authn_tokenizer_element_start
Used inside SamlAuthnRequestParser to help parse the AuthN request