Struct saml_rs::assertion::SubjectData

pub struct SubjectData {
    pub relay_state: String,
    pub qualifier: Option<BaseIDAbstractType>,
    pub qualifier_value: Option<String>,
    pub nameid_format: NameIdFormat,
    pub nameid_value: &'static str,
    pub acs: String,
    pub subject_not_on_or_after: DateTime<Utc>,
}

Data type for passing subject data in because yeaaaaah, specs

TODO: Justify the existence of the elements of this struct … more completely.

Fields
relay_state: String

Relay state as provided by the crate::AuthnRequest

qualifier: Option<BaseIDAbstractType>

Qualifier TODO: What’s the qualifier again?

qualifier_value: Option<String>

Qualifier value TODO: I really should know what these are

nameid_format: NameIdFormat

crate::sp::NameIdFormat, what kind of format you’re… going TODO oh no I’ve done it again

nameid_value: &'static str

NameID value - I know this one, it’s the reference to the user, like username or some rando noise if it’s transient. Regret, if it’s crate::sp::NameIdFormat::Kerberos

acs: String

The AssertionConsumerService - where we’ll send the request.

subject_not_on_or_after: DateTime<Utc>

The expiry of this Assertion. Woo, recovered there at the end.