Function saml_rs::cert::init_cert_from_base64

pub fn init_cert_from_base64(buf: &str) -> Result<X509, CertParseError>

this is a terrible function and only used for me to figure out how to parse out the cert from an SP Metadata file