Struct saml_rs::sp::ServiceProvider

pub struct ServiceProvider {
    pub entity_id: String,
    pub authn_requests_signed: bool,
    pub want_assertions_signed: bool,
    pub x509_certificate: Option<X509>,
    pub services: Vec<ServiceBinding>,
    pub protocol_support_enumeration: Option<String>,
    pub nameid_format: NameIdFormat,
}

SP metadata object, used for being able to find one when an AuthN request comes in (or IdP-initiated, eventually, maymbe?)

Fields
entity_id: String

EntityID

authn_requests_signed: bool

Will this SP send signed requests? If so, we’ll reject ones that aren’t.

want_assertions_signed: bool

Does this SP expect signed assertions?

x509_certificate: Option<X509>

The signing (public) certificate for the SP

services: Vec<ServiceBinding>

SP Services

protocol_support_enumeration: Option<String>

TODO protocol_support_enumeration? what’s this?

nameid_format: NameIdFormat

NameIdFormat - how we should identify the user

Implementations

impl ServiceProvider

pub fn test_generic(entity_id: &str) -> Self

Generate a test generic ServiceProvider with nonsense values for testing

fn service_attrib_parser(
&mut self,
servicetype: SamlBindingType,
attributes: Vec<OwnedAttribute>
) -> Result<ServiceBinding, String>

Used for handling the attributes of services in the tags of an SP Metadata XML file

Types tested (poorly):

AssertionConsumerService

SingleLogoutService

pub fn find_first_acs(&self) -> Result<ServiceBinding, &'static str>

return the first AssertionConsumerService we find

fn attrib_parser(&mut self, tag: &str, attributes: Vec<OwnedAttribute>)

Let’s parse some attributes!